Conditional display of token-request & vote bttons
If the user is on experiments/votes category they see the propose a vote
buttons vs if they're on the token requestes category they see the
propose a token request.

We're also now including the first post in the server payload, in html
format.

TODO: still can't get the modal to receive the passed error prop.